% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coral_reef_fishes_data.R
\docType{data}
\name{coral_reef_fishes_data}
\alias{coral_reef_fishes_data}
\title{coral_reef_fishes_data}
\format{An object of class \code{data.frame} with 6369 rows and 12 columns.}
\usage{
data(coral_reef_fishes_data)
}
\description{
coral_reef_fishes_data
}
\details{
This is the dataset provided by **add datapaper citation **
It contains the following data:
family: Family
genus: Genus
species: Species
id: Unique code identifying each individual
agei: Age i (years)
radi: Otolith radius at age i (mm)
agecap: Age of the fish at capture (years)
radcap: Radius of the otolith at capture (mm)
lencap: Length of the fish at capture (total length, mm)
l0p: Length of the fish at hatching (mm)
biomass: Wet body mass of the fish at capture (g)
location: Island of the sampling
}
\examples{
data(coral_reef_fishes_data)
}
\keyword{datasets}
