% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growthreg.R
\name{growthreg}
\alias{growthreg}
\title{A function to extract von Bertalanffy growth parameters for fishes from
back-calculated data}
\usage{
growthreg(
  length,
  age,
  id,
  lmax,
  linf_m = 0.8 * lmax,
  linf_sd = 0.2 * linf_m,
  l0_m,
  l0_sd = 0.2 * l0,
  plot = TRUE,
  ...
)
}
\arguments{
\item{length}{\code{\link[base]{numeric}} vector with length
(CAUTION: must be in cm).}

\item{age}{\code{\link[base]{numeric}} vector with age.}

\item{id}{\code{\link[base]{character}} vector with fish ID.}

\item{lmax}{Maximum size. Based on this value, maximum growth rate kmax
will be computed.}

\item{linf_m}{Prior for parameter linf, . (in cm)}

\item{linf_sd}{Prior sd for linf, asymptotic length (default set to 10\% of linf_m).}

\item{l0_m}{Prior for l0, size at hatching. (in cm)}

\item{l0_sd}{Prior sd for l0 (default set to 10\% of l0_m).}

\item{plot}{\code{\link[base]{logical}} Option to plot model fit. Defaults
to TRUE.}

\item{...}{Additional arguments to \code{\link[rstan]{sampling}}.}
}
\description{
A function to extract von Bertalanffy growth parameters for fishes from
back-calculated data
}
\details{
Returns a list with three elements.
The first element is a \code{\link[base]{data.frame}} with estimates for
\code{linf}, \code{k} and \code{t0}, \code{sl} and \code{gp}.
There is a hierarchical structure for \code{linf} and \code{k}, so that
there is a unique estimate for these parameters per individual
(\code{linf_j}, \code{k_j}).
\code{linf} and \code{k} are the population level estimates of \code{linf}
and \code{k}. \code{kmax} is the standardised growth parameter, depending
on \code{lmax} (kmax = exp(sl * log(lmax) + gp), see Morais and Bellwood
(2018) for details.
The second element is a \code{\link[base]{data.frame}} with model fits for
the average regression across individuals (\code{ypred_m}, \code{ypred_sd},
\code{ypred_lb}, \code{ypred_ub}), and the fitted regression er individual
(\code{yrep_m}, \code{yrep_sd}, \code{yrep_lb}, \code{yrep_ub}).
The third element is the entire stanfit object.
}
\examples{

\dontrun{
library(dplyr)
library(fishgrowbot)
data(coral_reef_fishes_data)
em <- filter(coral_reef_fishes_data,
             species == "Epinephelus merra",
             location == "Moorea")
bc <- bcalc(data = em)$lengths
growthreg(length = bc$l_m/10, age = bc$age, id = bc$id, lmax = 32,
          linf_m = 28, linf_sd = 5, l0_m = 0.15, l0_sd = 0.015,
          iter = 4000, chains = 1)
}
}
\author{
Nina M. D. Schiettekatte
}
