% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{bcplot}
\alias{bcplot}
\title{Function to plot results of back-calculation}
\usage{
bcplot(bcdata, colorid = FALSE, facet = FALSE, legend = FALSE, error = FALSE)
}
\arguments{
\item{bcdata}{Results from back-calculation (output of function bcalc())}

\item{colorid}{Option to specify colors by fish id (default = FALSE)}

\item{facet}{Option to split graph in facets per fish id (default = FALSE)}

\item{legend}{Option to add color legend (defaul = FALSE)}

\item{error}{Option to show 95\%CI around bc estimates (default = FALSE)}
}
\value{
A plot showing back-calculated lengths
}
\description{
Function to plot results of back-calculation
}
