% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{gmplot}
\alias{gmplot}
\title{Function to visualize the fitted hierarchical growth model}
\usage{
gmplot(gmdata, id = FALSE, facet = FALSE, legend = FALSE, error = TRUE)
}
\arguments{
\item{gmdata}{output of function growthreg()}

\item{id}{Show growth curves per individual (default = FALSE)}

\item{facet}{Option to split graph in facets per fish id (default = FALSE)}

\item{legend}{Option to add color legend (defaul = FALSE)}

\item{error}{Option to show 95\%CI around bc estimates (default = FALSE)}
}
\value{
A plot
}
\description{
Function to visualize the fitted hierarchical growth model
}
