% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growthreg.R
\name{growthreg}
\alias{growthreg}
\title{A function to extract growth parameters for fishes from back-calculation data}
\usage{
growthreg(length, age, id, lmax = 20, linf_m, plot = TRUE, ...)
}
\arguments{
\item{length}{Numerical vector with length (CAUTION: must be in cm)}

\item{age}{Numerical vector with age}

\item{id}{Character vector with fish id}

\item{lmax}{maximum size. Based on this value, maximum growth rate kmax will be computed.}

\item{linf_m}{Prior for linf}

\item{plot}{option to plot model fit (TRUE or FALSE)}

\item{...}{Additional arguments, see ?rstan::sampling()}
}
\description{
A function to extract growth parameters for fishes from back-calculation data
}
\details{
Returns a list with three elements.
First element is a dataframe with estimates for linf, k and t0, sl and gp.
There is a hierarchical structure for linf and k, so that there is a unique estimate for these parameters per individual (linf_j, k_j).
linf and k are the population level estimates of linf and k. kmax is the standardised growth parameter, depending on lmax
(kmax = exp(sl * log(lmax) + gp), see Morais and Bellwood (2018) for details)
Second element is a dataframe with model fits for the average regression across individuals (ypred_m, ypred_sd, ypred_lq, ypred_uq),
and the fitted regression er individual (urep_m, yrep_sd, yrep_lq, yrep_uq)
The third element is the entire stanfit object.
}
\examples{

em <- dplyr::filter(fishgrowbot::coral_reef_fishes_data, species == "Epinephelus merra", location == "Moorea")
bc <- fishgrowbot::bcalc(data = em)$lengths
fishgrowbot::growthreg(length = bc$l_m/10, age = bc$age,
id = bc$id, lmax = 32, linf_m = 30, iter = 2000, chains = 1)

}
\keyword{Bertalanfy}
\keyword{Von}
\keyword{fish,}
\keyword{growth,}
